refactor and fix existing unit tests
