+++ src/test/groovy/com/github/slugger/emby/sweeper/AppTest.groovy
@@ -1,6 +1,5 @@
 package com.github.slugger.emby.sweeper
 
-import groovyx.net.http.RESTClient
 import spock.lang.Specification
 import spock.lang.Unroll
 
@@ -42,44 +41,6 @@ class AppTest extends Specification {
             desc    | protocol  | input
             'set'   | 'https'   | true
             'unset' | 'http'    | false
-    }
-
-    def 'getUsers() returns an empty list when the specified user does not exist'() {
-        given:
-            def http = Mock(RESTClient) {
-                1 * get(path: '/Users') >> [data: [[Name: 'user1', Id: '1']]]
-                0 * _
-            }
-        expect:
-            new App(http: http, cleanupUsers: ['foo']).getUsers() == []
-    }
-
-    def 'getUsers() returns the proper subset of specified users'() {
-        given:
-            def findme = [Name: 'findme', Id: '2']
-            def http = Mock(RESTClient) {
-                1 * get(path: '/Users') >> [data: [[Name: 'user1', Id: '1'], findme, [Name: 'user2', Id: '3']]]
-                0 * _
-            }
-        expect:
-            new App(http: http, cleanupUsers: ['findme']).getUsers() == [findme]
-    }
-
-    @Unroll
-    def 'a user\'s views are filtered based on excluded lib options [#desc]'() {
-        given:
-            def user = [Id: '1']
-            def http = Mock(RESTClient) {
-                1 * get(path: "/Users/$user.Id/Views") >> [data: [Items: views]]
-            }
-        expect:
-            new App(http: http, excludedLibraries: excludes).getFilteredUserViews(user) == result
-        where:
-            desc | excludes    | views                                                     | result
-            1    | ['lib1']    | [[Name: 'lib1', Id: '10']]                                | []
-            2    | ['lib1']    | [[Name: 'lib2', Id: '20'], [Name: 'lib1', Id: '10']]      | ['20']
-            3    | []          | [[Name: 'lib2', Id: '20'], [Name: 'lib1', Id: '10']]      | ['20', '10']
-            4    | []          | []                                                        | []
     }
 
     def 'getApiKeyRequestHeaders() includes proper version info'() {
